Insignificant fixes

+++ app/src/main/java/com/example/formi/ulsukitchen/presentation/loot/LootFragment.java
@@ -2,6 +2,7 @@ package com.example.formi.ulsukitchen.presentation.loot;
 
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
@@ -9,18 +10,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.TextView;
 
 import com.arellomobile.mvp.MvpAppCompatFragment;
 import com.arellomobile.mvp.presenter.InjectPresenter;
-import com.example.formi.ulsukitchen.other.utils.TitleProvider;
 import com.example.formi.ulsukitchen.R;
 import com.example.formi.ulsukitchen.domain.dataclass.Eat;
-import com.example.formi.ulsukitchen.other.events.TitleEvent;
 import com.example.formi.ulsukitchen.other.itemdecorators.LinearItemDecorator;
-import com.example.formi.ulsukitchen.other.events.ShowBottomNavigationEvent;
+import com.example.formi.ulsukitchen.other.utils.TitleProvider;
 import com.example.formi.ulsukitchen.presentation.loot.adapter.LootAdapter;
-
-import org.greenrobot.eventbus.EventBus;
 
 import java.util.List;
 
@@ -36,6 +34,9 @@ public class LootFragment extends MvpAppCompatFragment implements LootView, Titl
     private LootAdapter adapter;
 
     private List<Eat> eatList;
+
+    private View emptyLoot;
+    private TextView txtAddEat;
 
     public static LootFragment newInstance(Bundle args) {
         LootFragment fragment = new LootFragment();
@@ -71,13 +72,36 @@ public class LootFragment extends MvpAppCompatFragment implements LootView, Titl
         @Override
         public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
             Eat eat = eatList.get(viewHolder.getAdapterPosition());
-            presenter.deleteItem(viewHolder.getAdapterPosition(), eat);
+            // TODO кастомизировать AlertDialog
+            AlertDialog.Builder ab = new AlertDialog.Builder(getActivity())
+                    .setTitle("Подтверждение действия")
+                    .setMessage("Вы действительно хотите удалить \"" + eat.title + "\" из корзины?")
+                    .setPositiveButton("Да", (dialog, which) -> {
+                        presenter.deleteItem(viewHolder.getAdapterPosition(), eat);
+                        dialog.dismiss();
+                    })
+                    .setNegativeButton("Нет", ((dialog, which) -> {
+                        adapter.notifyDataSetChanged();
+                        dialog.dismiss();
+                    }))
+                    .setCancelable(false);
+            AlertDialog alertDialog = ab.create();
+            alertDialog.setOnShowListener(dialog -> {
+                alertDialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(getResources().getColor(android.R.color.black));
+                alertDialog.getButton(AlertDialog.BUTTON_NEGATIVE).setTextColor(getResources().getColor(android.R.color.black));
+            });
+            alertDialog.show();
         }
     });
 
     @Override
     public void onViewCreated(final View view, final Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+
+        emptyLoot = view.findViewById(R.id.empty_loot);
+
+        txtAddEat = view.findViewById(R.id.txtAddToLoot);
+        txtAddEat.setOnClickListener((v) -> presenter.navigateToEat());
 
         recView = view.findViewById(R.id.recViewLoot);
         recView.setLayoutManager(new LinearLayoutManager(getActivity()));
@@ -99,8 +123,13 @@ public class LootFragment extends MvpAppCompatFragment implements LootView, Titl
         eatList.remove(position);
         adapter.notifyDataSetChanged();
         Snackbar.make(recView, title + " удален(а) из корзины", Snackbar.LENGTH_INDEFINITE)
-                .setAction("OK", (v) -> {})
+                .setAction("OK", (v) -> {
+                })
                 .show();
+        if (eatList.isEmpty()) {
+            emptyLoot.setVisibility(View.VISIBLE);
+            recView.setVisibility(View.GONE);
+        }
     }
 
     @Override
@@ -114,6 +143,13 @@ public class LootFragment extends MvpAppCompatFragment implements LootView, Titl
         this.eatList = eatList;
         adapter.setEatList(eatList);
         adapter.notifyDataSetChanged();
+        if (eatList.isEmpty()) {
+            emptyLoot.setVisibility(View.VISIBLE);
+            recView.setVisibility(View.GONE);
+        } else {
+            emptyLoot.setVisibility(View.GONE);
+            recView.setVisibility(View.VISIBLE);
+        }
     }
 
     @Override
